Move entities types to a separate file
