+++ exercises/07-normalizr/src/modules/entities/entities-slice.ts
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit'
 
-import { UserEntities } from './entities-schema'
+import { UserEntities } from './entities-types'
 import { usersActions } from 'modules/users/users-slice'
 
 const initSate: UserEntities = {
